step 1 of 3 for code refactoring at the server level
