--- triton-server/src/api/api-types.ts
@@ -1,6 +1,0 @@
-/* 
-    This is just a utility to make it easier to import the api types in the client project.
-    Client code should import types from here, rather than using a path to the external
-    types library, which could change.
-*/
-export * from "../../../triton-types/models/api"
