+++ triton-server/src/api/freezeman/authToken.ts
@@ -1,6 +1,6 @@
 import Axios, { AxiosResponse } from "axios"
 import config from "../../../config"
-import { logger } from "../../logger"
+import { logger } from "@core/logger"
 import jwtDecode from "jwt-decode"
 import { clearTimeout } from "timers"
