--- triton-server/src/download/api-types.ts
@@ -1,1 +1,0 @@
-export * from "../../../triton-types/models/api"
