--- triton-server/src/download/download-types.ts
@@ -1,2 +1,0 @@
-/* This utility makes the magic types available from the triton-types project */
-export * from "../../../triton-types/models/downloads"
