--- triton-server/src/freezeman/models.ts
@@ -1,1 +1,0 @@
-export * from "../../../triton-types/models/freezeman"
