+++ triton-server/src/notifications/contact-service.ts
@@ -3,13 +3,13 @@
  */
 /* eslint-disable no-console */
 
-import { Contact } from "./download/download-types"
-import config from "../config"
-import { ExternalProjectID } from "./api/api-types"
-import { logger } from "./logger"
-import { defaultDatabaseActions } from "./download/actions"
-import { sendEmail } from "./download/email"
-import { getProjectUsers } from "./magic/magic_api"
+import { Contact } from "../types/download"
+import config from "../../config"
+import { ExternalProjectID } from "../types/api"
+import { logger } from "@core/logger"
+import { defaultDatabaseActions } from "@database/download/actions"
+import { sendEmail } from "@notifications/emails/"
+import { getProjectUsers } from "@api/magic/magic_api"
 
 export function start() {
     logger.info("[contacts] Starting service...")
