+++ triton-server/src/types/api/index.ts
@@ -1,0 +1,19 @@
+/*
+    This is just a utility to make it easier to import the api types in the client project.
+    Client code should import types from here, rather than using a path to the external
+    types library, which could change.
+*/
+export {
+    ApiReply,
+    IsLoggedInData,
+    TritonDataset,
+    TritonDatasetFile,
+    TritonProject,
+    TritonReadset,
+    TritonRequest,
+    TritonRun,
+    User,
+    TritonCreateRequestResponse,
+    TritonCreateRequestBody,
+    ExternalProjectID,
+} from "../../../../triton-types/models/api"
