+++ triton-server/src/types/download/index.ts
@@ -1,0 +1,11 @@
+/* This utility makes the magic types available from the triton-types project */
+export {
+    Database,
+    NewDownloadFile,
+    DownloadFile,
+    NewDownloadRequest,
+    DownloadRequest,
+    DownloadRequestID,
+    DownloadDatasetID,
+    Contact,
+} from "../../../../triton-types/models/downloads"
