--- triton-server/src/types/index.ts
@@ -1,78 +1,0 @@
-import {
-    ApiReply,
-    IsLoggedInData,
-    TritonDataset,
-    TritonDatasetFile,
-    TritonProject,
-    TritonReadset,
-    TritonRequest,
-    TritonRun,
-    User,
-} from "../../../triton-types/models/api"
-import {
-    Database,
-    NewDownloadFile,
-    DownloadFile,
-    NewDownloadRequest,
-    DownloadRequest,
-    DownloadRequestID,
-    DownloadDatasetID,
-    Contact,
-} from "../../../triton-types/models/downloads"
-
-import {
-    Dataset,
-    DatasetFile,
-    DatasetFileID,
-    DatasetID,
-    DerivedSampleID,
-    FMSList,
-    Project,
-    ProjectID,
-    Readset,
-    ReleaseFlag,
-    ReleaseFlagAvailable,
-    ReleaseFlagBlocked,
-    ReleaseFlagReleased,
-    ValidationFlag,
-    ValidationFlagAvailable,
-    ValidationFlagFailed,
-    ValidationFlagPassed,
-} from "../../../triton-types/models/freezeman"
-
-export {
-    ApiReply,
-    IsLoggedInData,
-    TritonDataset,
-    TritonDatasetFile,
-    TritonProject,
-    TritonReadset,
-    TritonRequest,
-    TritonRun,
-    User,
-    Database,
-    NewDownloadFile,
-    DownloadFile,
-    NewDownloadRequest,
-    DownloadRequest,
-    DownloadRequestID,
-    DownloadDatasetID,
-    Contact,
-    Dataset,
-    DatasetFile,
-    DatasetFileID,
-    DatasetID,
-    DerivedSampleID,
-    FMSList,
-    Project,
-    ProjectID,
-    Readset,
-    ReleaseFlag,
-    ReleaseFlagAvailable,
-    ReleaseFlagBlocked,
-    ReleaseFlagReleased,
-    ValidationFlag,
-    ValidationFlagAvailable,
-    ValidationFlagFailed,
-    ValidationFlagPassed,
-}
