+++ triton-server/src/types/magic/index.ts
